+++ src/cypress/cypress.ts
@@ -1,3 +1,4 @@
 // imports all types here for commands
 import './types';
+import '@mmisty/cypress-tags';
 // any additional modules also