+++ tests/test-folder/mocha-events/failures/fail-one-test.test.ts
@@ -2,6 +2,7 @@ import {
   covergeAfterAllEvent,
   createResTest2,
   fixResult,
+  fullStepAttachment,
   whenCoverage,
   whenNoCoverage,
 } from '../../../cy-helper/utils';
@@ -75,19 +76,46 @@ describe('hello suite', () => {
     });
 
     it('check attachments', async () => {
-      expect(resFixed.map(t => t.attachments).sort()).toEqual([
-        [
-          {
-            name: 'hello suite -- hello test (failed).png',
-            source: 'source.png',
-            type: 'image/png',
-          },
-          {
-            name: 'test_0_number.cy.ts.mp4',
-            source: 'source.mp4',
-            type: 'video/mp4',
-          },
-        ],
+      expect(fullStepAttachment(resFixed, () => ({}))).toEqual([
+        {
+          attachments: [
+            {
+              name: 'hello suite -- hello test (failed).png',
+              source: 'source.png',
+              type: 'image/png',
+            },
+          ],
+          name: 'hello test',
+          parents: [
+            {
+              afters: [
+                {
+                  attachments: [
+                    {
+                      name: 'test_0_number.cy.ts.mp4',
+                      source: 'source.mp4',
+                      type: 'video/mp4',
+                    },
+                  ],
+                  name: 'video',
+                  status: 'passed',
+                  steps: [],
+                },
+              ],
+              befores: [
+                {
+                  attachments: [],
+                  name: '"before all" hook',
+                  status: 'passed',
+                  steps: [],
+                },
+              ],
+              suiteName: 'hello suite',
+            },
+          ],
+          status: 'failed',
+          steps: [],
+        },
       ]);
     });